+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Version 0.8 (2020-05-??)
 - feat, l4: add static html dashboard built with bokeh for postprocessing data
 - feat, www: create a new www folder that can be synced to s3 as our internal dashboard
 - feat, l4: add spike easing to confirmed cases, make spike easing more strict
+  - bugfix: found that some cases were yielding dips, so I added a workaround for those, but it's not a full fix though
 
 
 Version 0.7 (2020-05-06)